0.0.1 Initial Structures and Main Methods
Basic functionality of both Server and Client main methods complete. eg establishing connection, switching for valid commands

Planned structure is that each command will call a a method both on the client and the server.
These methods should be designed so that they interact in such a way so that they do not become out of sync.